Make World class control intro animation
